+++ src/Engine/ecs/Entity.cpp
@@ -1,0 +1,6 @@
+#include "stdafx.h"
+#include "Entity.h"
+
+namespace OKGame {
+    EntityId Entity::nextId = 0;
+}