+++ src/Engine/ecs/behaviors/EightDirectionBehavior.cpp
@@ -1,0 +1,6 @@
+#include "stdafx.h"
+#include "EightDirectionBehavior.h"
+
+namespace OKGame {
+    // Implementation in header
+}